requirements.txt
pandas
psycopg2-binary
shapely
config
dotenv
gevent
requests
socketIO-client
